Update README.md
This project implements a deep learning model for detecting whether a person is wearing a face mask or not. It utilizes Python and the Keras library with a pre-trained VGG16 architecture for training the model. The dataset used contains over 12,000 images of individuals with and without masks, preprocessed and split into training, testing, and validation sets. The model architecture is based on the VGG16 convolutional neural network, with additional custom layers for fine-tuning. Techniques such as transfer learning, Adam optimizer, ReLU activation function, and categorical cross-entropy loss are employed for efficient training. The trained model is tested on separate images to evaluate its performance, providing a reliable solution for face mask detection.